requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==3.2.0
bson==0.5.10
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
motor==3.6.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.18
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
